Common - proposal - Create proposal - Error message appears when user insert a valid amount #93 (#1033)

+++ src/Screens/Commons/CreateCommon/CreateStep2.js
@@ -58,7 +58,7 @@ const CreateStep2 = ({
   const [show, setShow] = useState(false);
 
   const minimumFieldRules = (currContribIndex) =>
-    `required|integer|min:${MIN_CONTRIBUTION[currContribIndex]}|max:${MAX_CONTRIBUTION[currContribIndex]}`;
+    `required|numeric|min:${MIN_CONTRIBUTION[currContribIndex]}|max:${MAX_CONTRIBUTION[currContribIndex]}`;
 
   useEffect(() => {
     fundingFormStore.registerFormField(
